hjkim_0621_1127 연차 일수 -> 분단위 변경

+++ src/main/java/knk/erp/api/shlee/account/dto/account/ReadDetail_AccountDTO.java
@@ -15,7 +15,7 @@ public class ReadDetail_AccountDTO {
     private String password;
     private String authority;
     private String phone;
-    private float vacation;
+    private int vacation;
     private String departmentName;
     private String address;
     private String email;
